Docs: Wrap setup shell command variables in quotes (#962)
In the setup, it is advised to store the desired site directory in a variable for later use, then the variable is used later for setting up paths. This would break the command if the path contains any spaces and is generally considered bad practice. Quoting your shell variables when using them (especially when paths are concerned) is [highly recommended][shellvars].

I can't remember if Windows acts the same.


[shellvars]: http://www.tldp.org/LDP/abs/html/quotingvar.html